Configure for automatically generating drafts by default
This is done to prevent a blog post from accidentally getting
published when it's not ready.

This change will cause the 'Status' in post metdata for all posts to
be considered 'draft' by default. To publish a post, it's status needs
to be explicitly set to 'published'.

As this change impacts older posts, they have been modified to
explicitly set the published status.