Add script to setup canonical redirection from old blog to new
The current blog is being hosted on github at 'naiquevin.github.io',
whereas the new blog will be located at 'www.naiquev.in'. To make sure
that old links are redirected to the new ones and that search engines
don't penalize the new blog due to duplicate content, the html files
on the old location will have to be changed to indicate canonical url
and redirect to it.

This script is written for this purpose. It will be used only once.