Fix styling for code blocks
- Use monospace font
- Fix margins and paddings
- Add horizontal scroll bar

Note that after this change, all posts need to be modified. This
commit adds a left margin for the code block. In all previous posts,
code inside the code blocks was indented by 4 spaces. This causes them
to shift more to the right. It will be fixed in a later commit.